requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
qdrant-client==1.7.0
pdfplumber==0.10.3
PyMuPDF==1.23.14
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
tiktoken==0.5.2
redis==5.0.1
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
